Move featureFlags action to featureFlags directory
