+++ src/app/featureFlags/index.ts
@@ -1,0 +1,9 @@
+import * as actions from './actions';
+import reducer from './reducer';
+import * as types from './types';
+
+export {
+  types,
+  reducer,
+  actions,
+};
